+++ src/lib/helpers/index.js
@@ -1,9 +1,7 @@
 import { handleMonophonicNoteTriggers } from "./handleMonophonicNoteTriggers";
 import { noteParser } from "./noteParser";
-import { setupComponent } from "./testHelpers";
 
 export {
   handleMonophonicNoteTriggers,
   noteParser,
-  setupComponent,
 };