+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.0.8
+- bundling with rollup
+- Chorus component
+
+## 0.0.2 - 0.0.7
+- Tremolo, Distortion, MonoSynth, FMsynth components
+- Connector component for in and output routing
+- Knob atom
+- AudioOutput component for audio output
+- MidiController component for handling midi input
+- Added tests
+
 ## 0.0.1
 
 - ComputerKeyboard component added